requirementsnew.txt
MarkupSafe==2.1.5
Pillow==8.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5-sip==12.17.0
PyYAML==6.0.2
QtPy==2.4.3
absl-py==2.3.1
certifi==2025.8.3
contourpy==1.3.0
cycler==0.11.0
filelock==3.13.1
fonttools==4.60.0
fsspec==2024.6.1
grpcio==1.75.0
h5py==2.10.0
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.4
kiwisolver==1.4.4
labelme==3.16.7
markdown==3.9
matplotlib==3.3.4
mkl-fft==1.3.11
mkl-random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
opencv-python==3.4.11.45
packaging==25.0
pip==25.2
protobuf==6.32.1
pyparsing==3.2.0
pyqt6==6.9.1
pyqt6-sip==13.10.2
pyreadline==2.1
python-dateutil==2.9.0.post0
scipy==1.5.4
setuptools==78.1.1
sip==6.12.0
six==1.17.0
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tomli==2.2.1
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tornado==6.5.1
tqdm==4.60.0
typing-extensions==4.12.2
werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
autocommand==2.2.2
backports.tarfile==1.2.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
platformdirs==4.2.2
typeguard==4.3.0
